Subcribes Companys로 위치변경
