+++ src/main/java/site/metacoding/miniproject2/domain/subcribes/SubcribesDao.java
@@ -2,7 +2,7 @@ package site.metacoding.miniproject2.domain.subcribes;
 
 import java.util.List;
 
-import site.metacoding.miniproject2.dto.SubribesRespDto.SubscribesListRespDto;
+import site.metacoding.miniproject2.dto.CompaysRespDto.SubscribesListRespDto;
 
 public interface SubcribesDao {
     // 이름 규칙에 맞는 메서드명으로 생성해주세요.
