+++ src/main/java/site/metacoding/miniproject2/dto/CompaysRespDto.java
@@ -1,5 +1,6 @@
 package site.metacoding.miniproject2.dto;
 
+import java.sql.Timestamp;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -75,6 +76,29 @@ public class CompaysRespDto {
         private Integer currentPage;
         private boolean isLast; // getter가 만들어지면 isLast() 이름으로 만들어짐. -> el에서는 last로 찾음
         private boolean isFirst; // getter가 만들어지면 isFirst() 이름으로 만들어짐. -> el에서는 first로 찾음
+    }
+
+    @Getter
+    @Setter
+    public static class SubscribesListRespDto {
+        private Integer id;
+        private Integer companysId;
+        private Integer wantedsId;
+        private String companyName;
+        private Timestamp created;// 공고시작시간
+        private Timestamp enddate;// 공고끝나는시간
+        private String state;
+
+        public SubscribesListRespDto(SubscribesListRespDto subcribes) {
+            this.id = id;
+            this.companysId = companysId;
+            this.wantedsId = wantedsId;
+            this.companyName = companyName;
+            this.created = created;
+            this.enddate = enddate;
+            this.state = state;
+        }
+
     }
 
 }
