+++ src/main/java/site/metacoding/miniproject2/service/WantedsService.java
@@ -58,7 +58,4 @@ public class WantedsService {
         return wantedList;
     }
 
-    public List<WantedListRespDto> findByIdCompanyId(Integer id) {
-        return null;
-    }
 }
