+++ src/main/java/site/metacoding/miniproject2/web/CompanysController.java
@@ -1,7 +1,10 @@
 package site.metacoding.miniproject2.web;
 
+import java.util.List;
+
 import javax.servlet.http.HttpSession;
 
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +19,7 @@ import site.metacoding.miniproject2.dto.CMRespDto;
 import site.metacoding.miniproject2.dto.CompaysReqDto.CompanysInsertReqDto;
 import site.metacoding.miniproject2.dto.CompaysReqDto.CompanysUpdateReqDto;
 import site.metacoding.miniproject2.dto.CompaysRespDto.CompanysInsertRespDto;
+import site.metacoding.miniproject2.dto.CompaysRespDto.SubscribesListRespDto;
 import site.metacoding.miniproject2.dto.SessionUsers;
 import site.metacoding.miniproject2.service.CompanysService;
 
@@ -45,7 +49,7 @@ public class CompanysController {
     @PutMapping("s/api/companys/{id}")
     public @ResponseBody CMRespDto<?> CompanyupdateId(@PathVariable Integer id,
             @RequestBody CompanysUpdateReqDto companysUpdateReqDto) {
-        companysService.Companyupdate(id, companysUpdateReqDto);
+        companysService.updateCompany(id, companysUpdateReqDto);
         return new CMRespDto<>(1, "회사정보수정성공", null);
     }
 
@@ -56,4 +60,10 @@ public class CompanysController {
         return new CMRespDto<>(1, "회사정보삭제", null);
     }
 
+    @GetMapping("s/api/subscribes/{id}")
+    public String subscribesform(@PathVariable Integer id, Model model) {
+        List<SubscribesListRespDto> subcribesList = companysService.subcribesListPage(id);
+        model.addAttribute("subcribesList", subcribesList);
+        return "subscribes/subscribes";
+    }
 }
