+++ src/main/resources/mapper/companys.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="site.metacoding.miniproject2.domain.companys.CompanysDao">
 
-	<select id="findByIdToDetail" resultType="site.metacoding.miniproject2.dto.CompaysRespDto">
+	<select id="findByIdToDetail" resultType="site.metacoding.miniproject2.dto.CompaysRespDto.CompaysRespDto">
 		SELECT c.id, c.company_name, c.address, c.email,
 		c.region_code_name, c.member_count, c.intro
 		FROM companys c
@@ -32,23 +32,11 @@
 		DELETE FROM companys WHERE id = #{id}
 	</delete>
 
-	<null
-		<null
-			<null
-				<null
-					<null
-						<null
-							<HEAD
-								<select id="findById" resultType="site.metacoding.miniproject2.dto$CompanysInsertRespDto">
-		select * from companys where id = #{id}
-	</select>
-								=======
-								<!-- 수현 작업함 -->
-								<update id="updateCompanysPhoto">
-		UPDATE companys SET
-		intro = #{intro}
-		where id=#{id}
-	</update>
-								<!-- 수현 작업완료 -->
-								>>>>>>> 443d20b14595a8b8c5190d1e00246cb54cedfdfd
+	<!-- 수현 작업함 -->
+	<update id="updateCompanysPhoto">
+			UPDATE companys SET
+			intro = #{intro}
+			where id=#{id}
+		</update>
+	<!-- 수현 작업완료 -->
 </mapper>
