+++ src/main/resources/mapper/subcribes.xml
@@ -2,6 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="site.metacoding.miniproject2.domain.subcribes.SubcribesDao">
 
+	<!--지원 작업-->
 	<select id="subcribesListPage"
 		resultType="site.metacoding.miniproject2.dto.SubribesRespDto">
 		SELECT w.company_id,w.* ,if(enddate>NOW(),'모집중','마감') state,
@@ -12,9 +13,10 @@
 		ON s.company_id= w.id
 		WHERE s.user_id = #{id};
 	</select>
+	<!--지원 작업 종료-->
 
 	<!-- 승현 작업함 -->
-	
+
 	<select id="findById"
 		resultType="site.metacoding.miniproject2.dto.SubribesRespDto.SubribesFindByIdRespDto">
 		SELECT * FROM subcribes WHERE id = #{id}
